Ensure that fireball is constantly checked to exist in the scene
In order to make sure that the fireball prefab
is constantly respawned in the scene after each deletion;
I check every frame to see if the cloned fireball is null
or not. If it is null, it is cloned frame fireballPrefab and
positioned in front of the enemyAI that aimlessly moves around.