Create a dummy script in order to attach to the character
The wandering AI enemy needs a way to recognize whether
or not the ray that is being cast from him has hit me or not.
In order to facilitate this process, I have created a dummy
script that I will attach to my character. The enemy can then
use this script to see whether its ray has indeed hit me or not.