Fix the bug related to the pos of the fireball
Initially, I thought that by finding the global pos of
the enemy and altering its z-coordinate; we would get the
ideal pos of where the fireball should appear in the scene
when it is cloned. However, this line of thought is flawed.
When we use transform.position property, we are working in
the global space. Changing this vector would result in a change
with respect to the global space again. This gives us the wrong
pos for the fireball. The solution is to first use the local
space and alter the vector that results from that, with respect
to the local space. Then change from local to global and set the
fireball's pos to that.