Check the plane by the z coordinate and not the y
The game objects created by unity have a local coordinate
system attached to their centers. When we do a rotation, this
local system moves as well. In my case, I rotated the object
by 90 degrees around the local x-axis. This caused the local
z-axis to switch places with the local y-axis. Therefore, we
ought to check the z-coordinate of the given point to determine
which plane we are on.