Check which plane the object is on by the y coordinate
The last commit is still correct. However I forgot the
fact that , based on my design, the points that are passed
to the function are in fact in the global space and not the
local one. Therefore, I changed it back to the original version
which checks the plane based on the y coordinate.