Change the iterator block name to that of the same as AITarget
The RayShooter class was going to notify FlyEnemy through die() that
it is dead. However, since both the AITarget and FlyEnemy share the
same die function, it would prevent repetition of if statements in
the RayShooter if these two functions had the same name. Therefore,
I chose the name Die() for both of them.