Create a messenger system to broadcast messages
At the moment, we would want the game objects to interact with the HUD
One way to do that would be to just reference the instances of scripts
attached to the game objects. However that would mean that the game and
HUD would be tightly coupled. That means that if in the future we wanted
to change something in the game, we would also need to consider how the
HUD should change as well as they influence each other.

In order to bypass this problem and make things as modular as possible,
I decided to use a messenger system written by someone else. Essentially
this system would enable us to register callback functions in the
messenger system and when an event occurs, it broadcasts messages to
these functions. The advantages of this over C# built-in event handling
system is that this system does not need to know where messages
originate from whereas C# needs to know the origin of such messages.