Add event listeners for score && broadcast message
We wanted to add score label bottom right of the screen that indicated
how many enemies in total we have killed from the start of the game.
Based on the addition of messenger system to out codebase, the easiest
way to handle this is to add a listener to UIController and broadcast
the message from the enemy source files after they are destroyed.