Keep track of number of enemies in the scene
In order to keep track of this number, I used the message system again.
In iterator blocks of the source files of the enemies, the moment they
are hit, a message is sent to the listener. However we need to make sure
that the messages are broadcasted before yield return statements. The
reason for this is that the effect of decrementing the number of enemies
is nullified by them being respawned. Since the comptutations of such
consecutive steps are done so blazingly fast, we wont see any change on
the screen.