Make sure spawned enemies have same speed as before
When we change the speed of the enemies in the scene via slider, their
speed does indeed change. However when they die, the newly spawned ones
have different speed than the destroyed ones. Therefore, I added an
event for such a situation and made sure that the speed of enemies in
EnemySpawn instance is replaced with whatever value the slider is given.